fix: better file allocation, errors fix
